Alternative mutation testing specification (#17)
* linear grading, line numbers not required

* bundle

* add prompt and shortname to MutantResult

* bundle

* tweaks

* bundle

* remove log, mutation breakpoint output text, bundle

* move **

* remove mutantAdvice in yml, combine points and total_faults into linearScoring in MutationTestUnit yml specification

* bundle, remove mutantAdvice usages
