+++ src/grading/builders/Builder.ts
@@ -19,6 +19,8 @@ export type MutantResult = {
   status: 'pass' | 'fail'
   tests: string[]
   output: string
+  prompt?: string
+  shortName?: string
   output_format?: OutputFormat
 }
